Fix #51 (talk back button not working)
