+++ app/src/androidTest/java/org/dharmaseed/android/ApplicationTest.java
@@ -17,18 +17,16 @@
  *
  */
 
-package org.dharmaseed.androidapp;
+package org.dharmaseed.android;
 
-import org.junit.Test;
-
-import static org.junit.Assert.*;
+import android.app.Application;
+import android.test.ApplicationTestCase;
 
 /**
- * To work on unit tests, switch the Test Artifact in the Build Variants view.
+ * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
  */
-public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() throws Exception {
-        assertEquals(4, 2 + 2);
+public class ApplicationTest extends ApplicationTestCase<Application> {
+    public ApplicationTest() {
+        super(Application.class);
     }
 }
