+++ app/src/main/java/org/dharmaseed/android/Center.java
@@ -1,4 +1,4 @@
-package org.dharmaseed.androidapp;
+package org.dharmaseed.android;
 
 import android.database.Cursor;
