+++ app/src/main/java/org/dharmaseed/android/CenterFetcherTask.java
@@ -17,32 +17,27 @@
  *
  */
 
-package org.dharmaseed.androidapp;
+package org.dharmaseed.android;
+
 
 /**
  * Created by bbethke on 2/19/16.
  */
-public class TalkFetcherTask extends DataFetcherTask {
+public class CenterFetcherTask extends DataFetcherTask {
 
-    public TalkFetcherTask(DBManager dbManager, NavigationActivity navigationActivity) {
+    public CenterFetcherTask(DBManager dbManager, NavigationActivity navigationActivity) {
         super(dbManager, navigationActivity);
     }
 
     @Override
     protected Void doInBackground(Void... params) {
 
-        updateTable(DBManager.C.Talk.TABLE_NAME, DBManager.C.Talk.ID,
-                "talks/",
+        updateTable(DBManager.C.Center.TABLE_NAME, DBManager.C.Center.ID,
+                "venues/",
                 new String[]{
-                        DBManager.C.Talk.TITLE,
-                        DBManager.C.Talk.DESCRIPTION,
-                        DBManager.C.Talk.VENUE_ID,
-                        DBManager.C.Talk.TEACHER_ID,
-                        DBManager.C.Talk.AUDIO_URL,
-                        DBManager.C.Talk.DURATION_IN_MINUTES,
-                        DBManager.C.Talk.UPDATE_DATE,
-                        DBManager.C.Talk.RECORDING_DATE,
-                        DBManager.C.Talk.RETREAT_ID
+                        DBManager.C.Center.WEBSITE,
+                        DBManager.C.Center.DESCRIPTION,
+                        DBManager.C.Center.NAME
                 });
 
         publishProgress();
