+++ app/src/main/java/org/dharmaseed/android/CenterRepository.java
@@ -1,4 +1,4 @@
-package org.dharmaseed.androidapp;
+package org.dharmaseed.android;
 
 import android.database.Cursor;
