+++ app/src/main/java/org/dharmaseed/android/DataFetcherTask.java
@@ -17,7 +17,7 @@
  *
  */
 
-package org.dharmaseed.androidapp;
+package org.dharmaseed.android;
 
 import android.content.ContentValues;
 import android.database.Cursor;
