+++ app/src/main/java/org/dharmaseed/android/DeleteTalkFragment.java
@@ -1,4 +1,4 @@
-package org.dharmaseed.androidapp;
+package org.dharmaseed.android;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
