+++ app/src/main/java/org/dharmaseed/android/StarCursorAdapter.java
@@ -17,7 +17,7 @@
  *
  */
 
-package org.dharmaseed.androidapp;
+package org.dharmaseed.android;
 
 
 import android.content.Context;
