+++ app/src/main/java/org/dharmaseed/android/TalkCursorAdapter.java
@@ -17,23 +17,25 @@
  *
  */
 
-package org.dharmaseed.androidapp;
+package org.dharmaseed.android;
 
 import android.content.Context;
 import android.database.Cursor;
+import android.graphics.BitmapFactory;
 import android.graphics.drawable.Drawable;
 import android.support.v4.content.ContextCompat;
-import android.support.v4.widget.CursorAdapter;
-import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-public class CenterCursorAdapter extends StarCursorAdapter {
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 
-    public CenterCursorAdapter(DBManager dbManager, NavigationActivity context, int layout, Cursor c) {
-        super(dbManager, DBManager.C.CenterStars.TABLE_NAME, context, layout, c);
+public class TalkCursorAdapter extends StarCursorAdapter {
+
+    public TalkCursorAdapter(DBManager dbManager, NavigationActivity context, int layout, Cursor c) {
+        super(dbManager, DBManager.C.TalkStars.TABLE_NAME, context, layout, c);
     }
 
     @Override
@@ -44,20 +46,25 @@ public class CenterCursorAdapter extends StarCursorAdapter {
     @Override
     public void bindView(View view, Context context, Cursor cursor) {
 
-        // Set center name
+        // Set talk title and teacher name
         TextView title=(TextView)view.findViewById(R.id.item_view_title);
-        TextView subtitle=(TextView)view.findViewById(R.id.item_view_subtitle);
-        title.setText(cursor.getString(cursor.getColumnIndexOrThrow(DBManager.C.Center.NAME)).trim());
-        subtitle.setText("");
+        TextView teacher=(TextView)view.findViewById(R.id.item_view_subtitle);
+        title.setText(cursor.getString(cursor.getColumnIndexOrThrow(DBManager.C.Talk.TITLE)).trim());
+        teacher.setText(cursor.getString(cursor.getColumnIndexOrThrow(DBManager.C.Teacher.NAME)).trim());
 
-        // Set photo
+        // Set teacher photo
+        String photoFilename = DBManager.getTeacherPhotoFilename(cursor.getInt(cursor.getColumnIndexOrThrow(DBManager.C.Talk.TEACHER_ID)));
         ImageView photoView = (ImageView) view.findViewById(R.id.item_view_photo);
-        Drawable icon = ContextCompat.getDrawable(context, R.drawable.dharmaseed_icon);
-        photoView.setImageDrawable(icon);
+        try {
+            FileInputStream photo = context.openFileInput(photoFilename);
+            photoView.setImageBitmap(BitmapFactory.decodeStream(photo));
+        } catch(FileNotFoundException e) {
+            Drawable icon = ContextCompat.getDrawable(context, R.drawable.dharmaseed_icon);
+            photoView.setImageDrawable(icon);
+        }
 
-        // Set stars
-        handleStars(view, cursor.getInt(cursor.getColumnIndexOrThrow(DBManager.C.Center.ID)));
-
+        // Set talk stars
+        handleStars(view, cursor.getInt(cursor.getColumnIndexOrThrow(DBManager.C.Talk.ID)));
     }
