+++ app/src/main/java/org/dharmaseed/android/TeacherFetcherTask.java
@@ -17,7 +17,7 @@
  *
  */
 
-package org.dharmaseed.androidapp;
+package org.dharmaseed.android;
 
 import android.content.Context;
 import android.database.Cursor;
