+++ app/src/main/java/org/dharmaseed/android/TeacherRepository.java
@@ -1,4 +1,4 @@
-package org.dharmaseed.androidapp;
+package org.dharmaseed.android;
 
 import android.database.Cursor;
 import android.util.Log;
