+++ app/src/main/res/layout/activity_play_talk.xml
@@ -28,7 +28,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="org.dharmaseed.androidapp.PlayTalkActivity">
+    tools:context="org.dharmaseed.android.PlayTalkActivity">
 
     <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
         android:orientation="horizontal"
