+++ app/src/main/res/layout/activity_retreat_search.xml
@@ -8,6 +8,6 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="org.dharmaseed.androidapp.RetreatSearchActivity">
+    tools:context="org.dharmaseed.android.RetreatSearchActivity">
 
 </LinearLayout>
