+++ app/src/main/res/layout/app_bar_navigation.xml
@@ -24,7 +24,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="org.dharmaseed.androidapp.NavigationActivity">
+    tools:context="org.dharmaseed.android.NavigationActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
