+++ app/src/test/java/org/dharmaseed/android/ExampleUnitTest.java
@@ -17,16 +17,18 @@
  *
  */
 
-package org.dharmaseed.androidapp;
+package org.dharmaseed.android;
 
-import android.app.Application;
-import android.test.ApplicationTestCase;
+import org.junit.Test;
+
+import static org.junit.Assert.*;
 
 /**
- * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
+ * To work on unit tests, switch the Test Artifact in the Build Variants view.
  */
-public class ApplicationTest extends ApplicationTestCase<Application> {
-    public ApplicationTest() {
-        super(Application.class);
+public class ExampleUnitTest {
+    @Test
+    public void addition_isCorrect() throws Exception {
+        assertEquals(4, 2 + 2);
     }
 }
